:ambulance: 사원이 휴가 신청 후, 관리자가 반려 했을 때 관리자 휴가 개수 느는 거 수정 / usedLeave 로직 수정

+++ src/main/java/com/semi/lynk/function/human/controller/EmployeeController.java
@@ -338,7 +338,7 @@ public class EmployeeController {
     @ResponseBody
     public Map<String, Object> cancelVacation(
             @RequestParam int draftNo,
-            @RequestParam int usedLeave,
+            @RequestParam float usedLeave,
             @RequestParam int employeeNo) {
 
         Map<String, Object> response = new HashMap<>();
